@@ -1,81 +1,56 @@
 {
-  "name": "tfgfront",
-  "version": "0.7.5",
   "private": true,
+  "name": "tfgfront",
+  "version": "0.8.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview --port 5173"
+  },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@loadable/component": "^5.15.3",
-    "@mui/icons-material": "^5.14.0",
-    "@mui/material": "^5.14.0",
-    "@mui/x-data-grid": "^6.10.2",
-    "@mui/x-date-pickers": "^6.10.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
-    "@types/loadable__component": "^5.13.4",
-    "@types/node": "^16.7.13",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@types/react-helmet": "^6.1.6",
-    "chart.js": "^4.4.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@loadable/component": "^5.16.4",
+    "@mui/icons-material": "^5.15.20",
+    "@mui/material": "^5.15.20",
+    "@mui/x-data-grid": "^7.7.0",
+    "@mui/x-date-pickers": "^7.7.0",
+    "@types/loadable__component": "^5.13.9",
+    "@types/node": "^20.14.2",
+    "@types/react-helmet": "^6.1.11",
+    "@types/react-transition-group": "^4.4.10",
+    "autoprefixer": "^10.4.19",
+    "chart.js": "^4.4.3",
     "class-validator": "^0.14.1",
-    "date-fns": "^2.30.0",
-    "dayjs": "^1.11.9",
-    "graphql": "^16.8.1",
-    "graphql-request": "^6.1.0",
+    "dayjs": "^1.11.11",
+    "graphql-request": "^7.0.1",
+    "postcss": "^8.4.38",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-intersection-observer": "^9.5.2",
     "react-query": "^3.39.3",
-    "react-responsive": "^9.0.2",
-    "react-router-dom": "^6.14.1",
-    "react-scripts": "5.0.1",
-    "react-toastify": "^9.1.3",
+    "react-responsive": "^10.0.0",
+    "react-router-dom": "^6.23.1",
+    "react-toastify": "^10.0.5",
     "react-transition-group": "^4.4.5",
-    "socket.io-client": "^4.7.1",
-    "tailwind-merge": "^1.14.0",
-    "typescript": "*",
-    "web-vitals": "^2.1.0",
+    "socket.io-client": "^4.7.5",
+    "tailwind-merge": "^2.3.0",
+    "tailwindcss": "^3.4.4",
     "zustand": "^4.5.2"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@typescript-eslint/eslint-plugin": "^5.61.0",
-    "@typescript-eslint/parser": "^5.61.0",
-    "eslint": "8.2.0",
-    "eslint-config-airbnb": "19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-standard-with-typescript": "^36.0.0",
-    "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "2.25.3",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "tailwindcss": "^3.3.2"
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.6",
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0"
   }
 }
