@@ -97,7 +97,7 @@ export function LibraryFilter(props:LibraryFilterProps):React.ReactElement {
                         <FormControl className="w-full">
                             <FormLabel>Dificultad</FormLabel>
                             <div className="flex gap-4 items-center">
-                                <Slider onChange={(e, v)=>{
+                                <Slider onChange={(_, v)=>{
                                     setDifficulty(v as number[]);
                                 }} value={difficulty} max={10} min={0} step={1}
                                 marks valueLabelDisplay="auto"
@@ -130,7 +130,7 @@ export function LibraryFilter(props:LibraryFilterProps):React.ReactElement {
                             <Autocomplete
                                 fullWidth
                                 value={genre}
-                                onChange={(e, v)=>{
+                                onChange={(_, v)=>{
                                     setGenre(v || "");
                                 }}
                                 renderInput={(params)=>(
@@ -148,7 +148,7 @@ export function LibraryFilter(props:LibraryFilterProps):React.ReactElement {
                             <Autocomplete
                                 fullWidth
                                 value={author}
-                                onChange={(e, v)=>{
+                                onChange={(_, v)=>{
                                     setAuthor(v || "");
                                 }}
                                 renderInput={(params)=>(
@@ -163,7 +163,7 @@ export function LibraryFilter(props:LibraryFilterProps):React.ReactElement {
                         </FormControl>
                         <FormControl className="py-4 w-full flex">
                             <FormControlLabel className="select-none" control={
-                                <Checkbox checked={readlist} onChange={(e, c)=>{
+                                <Checkbox checked={readlist} onChange={(_, c)=>{
                                     setReadlist(c);
                                 }}
                                 />
@@ -175,7 +175,7 @@ export function LibraryFilter(props:LibraryFilterProps):React.ReactElement {
                             <RadioGroup
                                 row
                                 value={readProgress}
-                                onChange={(e, v)=>{
+                                onChange={(_, v)=>{
                                     setReadProgress(v);
                                 }}
                             >
